# Make `tsc --help` 2x faster