### What does this PR do?

This adds a content-addressable runtime transpiler cache for input files larger than 50 KB. On macOS, these files are stored in `~/Library/Caches/bun/@t@/*.pile` (`t` is for "transpiler").

The goal is mostly for CLIs executed more than once, like `tsc`. Transpiler overhead is severe in cases like that.

macoS arm64:
```zig
Benchmark 1: bun ./node_modules/typescript/lib/tsc.js --help
  Time (mean ± σ):      84.1 ms ±   1.7 ms    [User: 73.7 ms, System: 10.2 ms]
  Range (min … max):    82.4 ms …  90.8 ms    34 runs

Benchmark 2: bun-1.0.13 ./node_modules/typescript/lib/tsc.js --help
  Time (mean ± σ):     205.8 ms ±   2.6 ms    [User: 181.6 ms, System: 23.9 ms]
  Range (min … max):   203.4 ms … 210.7 ms    13 runs

Benchmark 3: node ./node_modules/typescript/lib/tsc.js --help
  Time (mean ± σ):      94.5 ms ±   1.4 ms    [User: 83.9 ms, System: 8.7 ms]
  Range (min … max):    93.4 ms …  99.9 ms    30 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Summary
  bun ./node_modules/typescript/lib/tsc.js --help ran
    1.12 ± 0.03 times faster than node ./node_modules/typescript/lib/tsc.js --help
    2.45 ± 0.06 times faster than bun-1.0.13 ./node_modules/typescript/lib/tsc.js --help
```

`vite --help` on macOS arm64:

```zig
Benchmark 1: bun --bun ./node_modules/.bin/vite --version
  Time (mean ± σ):      94.2 ms ±   1.1 ms    [User: 89.0 ms, System: 17.5 ms]
  Range (min … max):    91.8 ms …  96.0 ms    29 runs

Benchmark 2: bun-1.0.13 --bun ./node_modules/.bin/vite --version
  Time (mean ± σ):     117.6 ms ±   1.5 ms    [User: 164.2 ms, System: 26.2 ms]
  Range (min … max):   115.5 ms … 121.9 ms    23 runs

Benchmark 3: node ./node_modules/.bin/vite --version
  Time (mean ± σ):      99.9 ms ±   1.4 ms    [User: 95.5 ms, System: 10.5 ms]
  Range (min … max):    98.0 ms … 103.3 ms    27 runs

Summary
  bun --bun ./node_modules/.bin/vite --version ran
    1.06 ± 0.02 times faster than node ./node_modules/.bin/vite --version
    1.25 ± 0.02 times faster than bun-1.0.13 --bun ./node_modules/.bin/vite --version
```

### How did you verify your code works?

This needs tests. I didn't write them yet.